Further cleanup of code. Removes dependencies & simplifies IncitoLoader model
