--- Cartfile
@@ -1,3 +1,0 @@
-
-github "shopgun/swift-generic-geometry" ~> 0.3
-github "shopgun/swift-future" ~> 0.4
