+++ Incito.podspec
@@ -1,7 +1,7 @@
 Pod::Spec.new do |s|
 
     s.name            = "Incito"
-    s.version         = "0.10"
+    s.version         = "1.0"
     s.summary         = "Incito viewer for iOS."
     s.description     = <<-DESC
                          A library for loading and viewing Incito-format documents.
@@ -11,7 +11,7 @@ Pod::Spec.new do |s|
     s.author           = "ShopGun"
     s.social_media_url = "http://twitter.com/ShopGun"
 
-    s.platform         = :ios, "9.3"
+    s.platform         = :ios, "11.0"
     s.swift_version    = "5.0.1"
     s.pod_target_xcconfig = { 'SWIFT_VERSION' => '5.0' }
 
@@ -19,7 +19,4 @@ Pod::Spec.new do |s|
     
     s.source_files = "Sources/**/*.{swift,h,m}"
     s.resources = ["Sources/**/Resources/**/*.strings"]
-    
-    s.dependency "ShopGun-GenericGeometry", "~> 0.3"
-    s.dependency "ShopGun-Future", "~> 0.4"
 end
