+++ Sources/Resources/sv.lproj/Incito.strings
