--- Tests/LayoutTests/incito-blocktest-375.dimensions.json
@@ -1,114 +1,0 @@
-{
-  "id": "a",
-  "rect": {
-    "x": 0,
-    "y": 0,
-    "width": 375,
-    "height": 398
-  },
-  "children": [
-    {
-      "id": "b",
-      "rect": {
-        "x": 0,
-        "y": 0,
-        "width": 375,
-        "height": 398
-      },
-      "children": [
-        {
-          "id": "c",
-          "rect": {
-            "x": 7,
-            "y": 7,
-            "width": 361,
-            "height": 100
-          },
-          "children": [
-            {
-              "id": "d",
-              "rect": {
-                "x": 9,
-                "y": 9,
-                "width": 30,
-                "height": 50
-              },
-              "children": []
-            },
-            {
-              "id": "e",
-              "rect": {
-                "x": 11,
-                "y": 63,
-                "width": 100,
-                "height": 10
-              },
-              "children": []
-            },
-            {
-              "id": "f",
-              "rect": {
-                "x": 13,
-                "y": 79,
-                "width": 361,
-                "height": 10
-              },
-              "children": []
-            }
-          ]
-        },
-        {
-          "id": "g",
-          "rect": {
-            "x": 9,
-            "y": 111,
-            "width": 200,
-            "height": 200
-          },
-          "children": [
-            {
-              "id": "h",
-              "rect": {
-                "x": 29,
-                "y": 131,
-                "width": 20,
-                "height": 20
-              },
-              "children": []
-            },
-            {
-              "id": "i",
-              "rect": {
-                "x": 13,
-                "y": 175,
-                "width": 100,
-                "height": 20
-              },
-              "children": []
-            },
-            {
-              "id": "j",
-              "rect": {
-                "x": 15,
-                "y": 217,
-                "width": 200,
-                "height": 20
-              },
-              "children": []
-            }
-          ]
-        },
-        {
-          "id": "k",
-          "rect": {
-            "x": 11,
-            "y": 317,
-            "width": 365,
-            "height": 70
-          },
-          "children": []
-        }
-      ]
-    }
-  ]
-}
