--- Tests/LayoutTests/incito-flextest-375.dimensions.json
@@ -1,155 +1,0 @@
-{
-  "id": "a",
-  "rect": {
-    "x": 0,
-    "y": 0,
-    "width": 375,
-    "height": 400
-  },
-  "children": [
-    {
-      "id": "b",
-      "rect": {
-        "x": 12.5,
-        "y": 0,
-        "width": 350,
-        "height": 200
-      },
-      "children": [
-        {
-          "id": "c",
-          "rect": {
-            "x": 22.5,
-            "y": 10,
-            "width": 100,
-            "height": 30
-          },
-          "children": []
-        },
-        {
-          "id": "d",
-          "rect": {
-            "x": 22.5,
-            "y": 40,
-            "width": 330,
-            "height": 90
-          },
-          "children": [
-            {
-              "id": "e",
-              "rect": {
-                "x": 22.5,
-                "y": 40,
-                "width": 30,
-                "height": 50
-              },
-              "children": []
-            },
-            {
-              "id": "f",
-              "rect": {
-                "x": 22.5,
-                "y": 90,
-                "width": 165,
-                "height": 27
-              },
-              "children": []
-            },
-            {
-              "id": "g",
-              "rect": {
-                "x": 22.5,
-                "y": 117,
-                "width": 80,
-                "height": 20
-              },
-              "children": []
-            }
-          ]
-        },
-        {
-          "id": "h",
-          "rect": {
-            "x": 22.5,
-            "y": 130,
-            "width": 300,
-            "height": 60
-          },
-          "children": []
-        }
-      ]
-    },
-    {
-      "id": "i",
-      "rect": {
-        "x": 0,
-        "y": 200,
-        "width": 375,
-        "height": 200
-      },
-      "children": [
-        {
-          "id": "j",
-          "rect": {
-            "x": 20,
-            "y": 220,
-            "width": 52.5,
-            "height": 160
-          },
-          "children": []
-        },
-        {
-          "id": "k",
-          "rect": {
-            "x": 82.5,
-            "y": 210,
-            "width": 182.5,
-            "height": 180
-          },
-          "children": [
-            {
-              "id": "l",
-              "rect": {
-                "x": 82.5,
-                "y": 210,
-                "width": 30,
-                "height": 50
-              },
-              "children": []
-            },
-            {
-              "id": "m",
-              "rect": {
-                "x": 82.5,
-                "y": 260,
-                "width": 91.25,
-                "height": 54
-              },
-              "children": []
-            },
-            {
-              "id": "n",
-              "rect": {
-                "x": 82.5,
-                "y": 314,
-                "width": 80,
-                "height": 20
-              },
-              "children": []
-            }
-          ]
-        },
-        {
-          "id": "o",
-          "rect": {
-            "x": 265,
-            "y": 210,
-            "width": 100,
-            "height": 100
-          },
-          "children": []
-        }
-      ]
-    }
-  ]
-}
