--- Tests/LayoutTests/incito-stroketest-375.dimensions.json
@@ -1,32 +1,0 @@
-{
-  "id": "a",
-  "rect": {
-    "x": 0,
-    "y": 0,
-    "width": 375,
-    "height": 225
-  },
-  "children": [
-    {
-      "id": "b",
-      "rect": {
-        "x": 50,
-        "y": 50,
-        "width": 275,
-        "height": 125
-      },
-      "children": [
-        {
-          "id": "c",
-          "rect": {
-            "x": 55,
-            "y": 60,
-            "width": 100,
-            "height": 100
-          },
-          "children": []
-        }
-      ]
-    }
-  ]
-}
