--- Tests/LayoutTests/incito-stroketest-375.json
@@ -1,43 +1,0 @@
-{
-    "version": "1.0.0",
-    "id": "stroke-test-375",
-    "theme": {
-        "background_color": "#5c5c5c"
-    },
-    "root_view": {
-        "id": "a",
-        "layout_width": 375,
-        "padding": 50,
-        "gravity": "center_horizontal",
-        "background_color": "#fafafa",
-        "child_views": [
-            {
-                "id": "b",
-                "padding_bottom": 10,
-                "padding_top": 5,
-                "stroke_width": 5,
-                "stroke_style": "dashed",
-                "stroke_color": "rgba(100%, 100%, 0%, 20%)",
-                "background_color": "rgba(100%, 0%, 0%, 20%)",
-
-                "child_views": [
-                    {
-                        "id": "c",
-                        "layout_height": 100,
-                        "layout_width": 100,
-                        "padding_bottom": 10,
-
-                        "background_color": "rgba(90%, 90%, 90%, 100%)",
-                        "stroke_width": 10,
-                        "stroke_bottom_width": 20,
-                        "stroke_style": "solid",
-                        "stroke_color": "#960911",
-                        "stroke_bottom_color": "#ff0911",
-
-                        "corner_radius": 20
-                    }
-                ]
-            }
-        ]
-    }
-}
