--- Tests/LayoutTests/incito-transformtest-350.dimensions.json
@@ -1,104 +1,0 @@
-{
-  "id": null,
-  "rect": {
-    "x": 0,
-    "y": 0,
-    "width": 350,
-    "height": 1500
-  },
-  "children": [
-    {
-      "id": null,
-      "rect": {
-        "x": 10.39764404296875,
-        "y": 23.570575714111328,
-        "width": 198.38067626953125,
-        "height": 253.80003356933594
-      },
-      "children": [
-        {
-          "id": null,
-          "rect": {
-            "x": 72.255859375,
-            "y": 35.36619186401367,
-            "width": 130.26593017578125,
-            "height": 110.4576416015625
-          },
-          "children": [
-            {
-              "id": null,
-              "rect": {
-                "x": 77.652099609375,
-                "y": 62.15610885620117,
-                "width": 123.222412109375,
-                "height": 73.56666564941406
-              },
-              "children": []
-            }
-          ]
-        }
-      ]
-    },
-    {
-      "id": null,
-      "rect": {
-        "x": 47,
-        "y": 47,
-        "width": 6,
-        "height": 6
-      },
-      "children": []
-    },
-    {
-      "id": null,
-      "rect": {
-        "x": 50,
-        "y": 281.5959777832031,
-        "width": 290.5445556640625,
-        "height": 350.3117980957031
-      },
-      "children": [
-        {
-          "id": null,
-          "rect": {
-            "x": 75.63427734375,
-            "y": 353.58648681640625,
-            "width": 97.7264404296875,
-            "height": 32.275634765625
-          },
-          "children": []
-        },
-        {
-          "id": null,
-          "rect": {
-            "x": 114.9493408203125,
-            "y": 370.4521789550781,
-            "width": 126.44384765625,
-            "height": 78.630126953125
-          },
-          "children": []
-        },
-        {
-          "id": null,
-          "rect": {
-            "x": 135.47052001953125,
-            "y": 462.6468200683594,
-            "width": 61.51483154296875,
-            "height": 25.153228759765625
-          },
-          "children": []
-        },
-        {
-          "id": null,
-          "rect": {
-            "x": 155.99169921875,
-            "y": 483.2153015136719,
-            "width": 126.4439697265625,
-            "height": 78.63009643554688
-          },
-          "children": []
-        }
-      ]
-    }
-  ]
-}
